Гипотетически, такой скрипт должен работать

+++ tv_tuner_channels_switcher.sh
@@ -2,29 +2,37 @@
 MY_DIR=`dirname $0`
 $MY_DIR/channels_arrays.sh
 
+curStandard=""
+
 while true
 do
-    echo 'Надо ввести номер канала'
-    read chanel
-    case "$chanel" in
-        11)
-            freq="151.25"
-            standard="pal-D"
-            ;;
-        56)
-            freq="511.25"
-            standard="secam-D"
-            ;;
-        "exit")
-            killall -9 vlc
-            exit 0
-            ;;
-        *)
-            echo 'неправильный номер'
-            ;;
-    esac
-    killall -9 vlc >/dev/null 2>&1
-    # Delay
-    sleep 0.2 >/dev/null 2>&1
-    v4l2-ctl -s "$standard"; ivtv-tune -f "$freq"; vlc -vvv pvr:// :pvr-device="/dev/video0" :pvr-radio-device="/dev/radio0" --deinterlace=1 >/dev/null 2>&1 &
+    echo "Введите номер канала"
+    read channelNumber
+
+    if [ "$channelNumber" == "exit" ]
+    then
+        killall -9 vlc
+        exit 0
+    fi
+
+    freq=${chFreq[$channelNumber]}
+    standard=${chStandard[$channelNumber]}
+    name=${chName[$channelNumber]}
+
+    if [ -z "$freq" ]
+    then
+        echo "Такой канал не определён"
+    else
+        echo "Переключение на канал №$channelNumber. Название: $name; Частота: $freq; стандарт: $standard"
+        if [ "$curStandard" == "$standard" ]
+        then
+            ivtv-tune -f "$freq" >/dev/null 2>&1
+        else
+            killall -9 vlc >/dev/null 2>&1
+            sleep 0.2 >/dev/null 2>&1
+            v4l2-ctl -s "$standard" >/dev/null 2>&1
+            ivtv-tune -f "$freq" >/dev/null 2>&1
+            vlc -vvv pvr:// :pvr-device="/dev/video0" :pvr-radio-device="/dev/radio0" --deinterlace=1 >/dev/null 2>&1 &
+        fi
+    fi
 done
